Switched away from ad-hoc parser to PEG.js encoded parser for handling more complex recipes.
The old parser was a hack.  This uses a parser expression grammar to allow for facile changes to the recipe specification language.  The parser grammar is
encoded in a weird pseudo-javascript DSL that's currently included in the parser.coffee file and the parser is generated at runtime.  This could be done offline server-side and included manually but this is simpler for rapid development.